Fixed scroll bug
After changing the #history-wrapper to 'justify-content: flex-end;' the y-scrolling was broken. This appears to be due to a bug in flexbox.

Added a JS line after the calculation of the input to scroll to the bottom, now 'justify-content: flex-start;' can be used and the new result will be visible regardless.